http_/fastapi_proj1/pyproject.toml
[tool.poetry]
name = "fastapi-proj1"
version = "0.1.0"
description = ""
authors = ["nur950960 <yanshanlo2006@gamil.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.110.0"
unicorn = {extras = ["standard"], version = "^2.0.1.post1"}
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.29"
python-decouple = "^3.8"
uvicorn = {extras = ["standard"], version = "^0.29.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
